+++ discord/channel.py
@@ -147,7 +147,7 @@ class ForumTag(Hashable):
         moderated: :class:`bool`
             Whether this tag can only be added or removed by a moderator with
             the :attr:`~Permissions.manage_threads` permission.
-        emoji: :class:`PartialEmoji`
+        emoji: Optional[:class:`PartialEmoji`]
             The emoji that is used to represent this tag.
             Note that if the emoji is a custom emoji, it will *not* have name information.
     """
@@ -155,19 +155,19 @@ class ForumTag(Hashable):
     __slots__ = ("name", "id", "moderated", "emoji")
 
     def __init__(
-        self, *, name: str, emoji: EmojiInputType, moderated: bool = False
+        self, *, name: str, emoji: EmojiInputType | None = None, moderated: bool = False
     ) -> None:
         self.name: str = name
         self.id: int = 0
         self.moderated: bool = moderated
-        self.emoji: PartialEmoji
+        self.emoji: PartialEmoji | None = None
         if isinstance(emoji, _EmojiTag):
             self.emoji = emoji._to_partial()
         elif isinstance(emoji, str):
             self.emoji = PartialEmoji.from_str(emoji)
-        else:
+        elif emoji is not None:
             raise TypeError(
-                "emoji must be a GuildEmoji, PartialEmoji, or str and not"
+                "emoji must be a GuildEmoji, PartialEmoji, str, or None and not"
                 f" {emoji.__class__!r}"
             )
 
@@ -189,14 +189,23 @@ class ForumTag(Hashable):
 
         emoji_name = data["emoji_name"] or ""
         emoji_id = utils._get_as_snowflake(data, "emoji_id") or None
-        self.emoji = PartialEmoji.with_state(state=state, name=emoji_name, id=emoji_id)
+        if not emoji_name and not emoji_id:
+            self.emoji = None
+        else:
+            self.emoji = PartialEmoji.with_state(
+                state=state, name=emoji_name, id=emoji_id
+            )
         return self
 
     def to_dict(self) -> dict[str, Any]:
         payload: dict[str, Any] = {
             "name": self.name,
             "moderated": self.moderated,
-        } | self.emoji._to_forum_reaction_payload()
+        }
+        if self.emoji is not None:
+            payload |= self.emoji._to_forum_reaction_payload()
+        else:
+            payload |= {"emoji_id": None, "emoji_name": None}
 
         if self.id:
             payload["id"] = self.id