+++ CHANGELOG.md
@@ -18,6 +18,8 @@ These changes are available on the `master` branch, but have not yet been releas
 
 - Include `bypass_slowmode` in `Permissions.all`.
   ([#3231](https://github.com/Pycord-Development/pycord/pull/3231))
+- Allow `ForumTag` to be created without an emoji.
+  ([#3245](https://github.com/Pycord-Development/pycord/pull/3245))
 
 ### Deprecated